관리자가 상품 삭제하는 API 구현

+++ shoppingmall-backend/src/main/java/com/cafe24/shoppingmall/service/ProductService.java
@@ -132,7 +132,16 @@ public class ProductService {
 		if(!categoryDao.deleteProductCategories(no)) {
 			return false;
 		}
+		if(!productDao.deleteImages(no)) {
+			return false;
+		}
+		if(!categoryDao.deleteProductCategories(no)) {
+			return false;
+		}
 		if(!bucketItemDao.deleteItemsByProductNo(no)) {
+			return false;
+		}
+		if(!ordersDao.updateProductOptionItemNoToNull(no)) {
 			return false;
 		}
 		if(!productOptionDao.deleteOptions(no)) {
